lazySplittable refactored to ShallowSplittable making way for lazier index/key lookup
